:recycle: Push YG employee sync to a standardized service.
