+++ api/src/services/index.ts
@@ -24,3 +24,4 @@ export * as TravelDeskTravelAgencies from "./travel-desk-travel-agencies"
 export * as TravelDeskTravelRequests from "./travel-desk-travel-requests"
 export * as TravelSegments from "./travel-segments"
 export * as Users from "./users"
+export * as YgEmployees from "./yg-employees"
