+++ api/src/services/yg-employees/index.ts
@@ -1,0 +1,1 @@
+export { SyncService } from "./sync-service"
